chore: rename foxyBalancesApi -> foxyApi
