+++ src/state/store.ts
@@ -5,7 +5,7 @@ import { PERSIST, persistReducer, persistStore } from 'redux-persist'
 import { getStateWith, registerSelectors } from 'reselect-tools'
 import { swapperApi } from 'state/apis/swapper/swapperApi'
 
-import { foxyBalancesApi } from './apis/foxy/foxyBalancesApi'
+import { foxyApi } from './apis/foxy/foxyApi'
 import { apiSlices, reducer, ReduxState, slices } from './reducer'
 import { assetApi } from './slices/assetsSlice/assetsSlice'
 import { marketApi, marketData } from './slices/marketDataSlice/marketDataSlice'
@@ -26,7 +26,7 @@ const apiMiddleware = [
   assetApi.middleware,
   txHistoryApi.middleware,
   validatorDataApi.middleware,
-  foxyBalancesApi.middleware,
+  foxyApi.middleware,
   swapperApi.middleware,
 ]
